weapon Float

+++ division2calc/weapon/__init__.py
@@ -4,6 +4,7 @@ from typing import Literal
 import division2calc.weapon.attrs as attrs
 import division2calc.weapon.mods as mods
 import division2calc.weapon.talents as talents
+from division2calc.utils import Float
 
 WeaponType = Literal['AR', 'SMG', 'LMG', 'Rifle', 'MMR', 'Shotgun']
 
@@ -32,65 +33,65 @@ class Weapon:
         self.mods = (self.optics, self.magazine, self.muzzle, self.underbarrel)
 
     @property
-    def weapon_damage(self) -> float:
-        return self.expertise_level * 0.01
+    def weapon_damage(self) -> Float:
+        return Float(self.expertise_level * 0.01)
 
     @property
-    def weapon_type_damage(self) -> float:
+    def weapon_type_damage(self) -> Float:
         pct = 0
         # cores
         for core in self.cores:
             if self.type == 'AR' and isinstance(core, attrs.AssultRifleDamage):
                 pct += core.pct
         # result
-        return pct
+        return Float(pct)
 
     @property
-    def critical_hit_chance(self) -> float:
+    def critical_hit_chance(self) -> Float:
         pct = 0
         # mods
         for mod in self.mods:
             if isinstance(mod, mods.CriticalHitChance):
                 pct += mod.pct
         # result
-        return pct
+        return Float(pct)
 
     @property
-    def critical_hit_damage(self) -> float:
+    def critical_hit_damage(self) -> Float:
         pct = 0
         # mods
         for mod in self.mods:
             if isinstance(mod, mods.CriticalHitDamage):
                 pct += mod.pct
         # result
-        return pct
+        return Float(pct)
 
     @property
-    def headshot_damage(self) -> float:
+    def headshot_damage(self) -> Float:
         pct = 0
         # mods
         for mod in self.mods:
             if isinstance(mod, mods.HeadshotDamage):
                 pct += mod.pct
         # result
-        return pct
+        return Float(pct)
 
     @property
-    def damage_to_health(self) -> float:
+    def damage_to_health(self) -> Float:
         pct = 0
         # attr
         for attr in self.attrs:
             if isinstance(attr, attrs.HealthDamage):
                 pct += attr.pct
         # result
-        return pct
+        return Float(pct)
 
     @property
-    def damage_to_target_out_of_cover(self) -> float:
+    def damage_to_target_out_of_cover(self) -> Float:
         pct = 0
         # attr
         for attr in self.attrs:
             if isinstance(attr, attrs.DamageToTargetOutOfCover):
                 pct += attr.pct
         # result
-        return pct
+        return Float(pct)
